+++ examples/websocket/CombinedDepth/CombinedDepth.go
@@ -8,18 +8,18 @@ import (
 )
 
 func main() {
-	WsDepthHandlerExample()
+	WsCombinedDepthHandlerExample()
 }
 
-func WsDepthHandlerExample() {
-	websocketStreamClient := binance_connector.NewWebsocketStreamClient(false)
-	wsDepthHandler := func(event *binance_connector.WsDepthEvent) {
+func WsCombinedDepthHandlerExample() {
+	websocketStreamClient := binance_connector.NewWebsocketStreamClient(true)
+	wsCombinedDepthHandler := func(event *binance_connector.WsDepthEvent) {
 		fmt.Println(binance_connector.PrettyPrint(event))
 	}
 	errHandler := func(err error) {
 		fmt.Println(err)
 	}
-	doneCh, stopCh, err := websocketStreamClient.WsDepthServe("LTCBTC", wsDepthHandler, errHandler)
+	doneCh, stopCh, err := websocketStreamClient.WsCombinedDepthServe([]string{"LTCBTC", "BTCUSDT", "MATICUSDT"}, wsCombinedDepthHandler, errHandler)
 	if err != nil {
 		fmt.Println(err)
 		return