+++ examples/websocket/aggtrades/aggtrades.go
@@ -11,13 +11,16 @@ func main() {
 }
 
 func AggTradesExample() {
+	// Initialise Websocket Client with Testnet base URL and false for "isCombined" parameter
+	websocketStreamClient := binance_connector.NewWebsocketStreamClient(false, "wss://testnet.binance.vision")
+
 	wsAggTradeHandler := func(event *binance_connector.WsAggTradeEvent) {
 		fmt.Println(binance_connector.PrettyPrint(event))
 	}
 	errHandler := func(err error) {
 		fmt.Println(err)
 	}
-	doneCh, _, err := binance_connector.WsAggTradeServe("LTCBTC", wsAggTradeHandler, errHandler)
+	doneCh, _, err := websocketStreamClient.WsAggTradeServe("BTCUSDT", wsAggTradeHandler, errHandler)
 	if err != nil {
 		fmt.Println(err)
 		return